moves xml decoder to xml package
